Minor change in structure of inner classes
